+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+- New Roblox Badges feature
+
 - Fixed settings header layout
 - Fixed dropdown button appearing in other dropdowns
 - Fixed dropdown button for users having roblox website under parental controls
 - Dropdown button is now 10th/9th in the list
 - Fixed profile picture animation flickering after trying to replay it
 - Fixed presence status icons in group members viewer
+- Fixed memory leak on settings page